# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "myExtBot API",
    "version": "0.1.0",
    "description": "REST API for myExtBot — an extensible multi-agent orchestration platform.\n\nAll endpoints under `/api/*` require an `x-api-key` header (set `MYEXTBOT_API_KEY` env var, default: `dev-key`).\n\nInteractive docs: `/api-docs` · Raw spec: `/api-docs/json`"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "AgentSla": {
        "type": "object",
        "properties": {
          "maxCostPerCall": {
            "type": "number"
          },
          "maxCallsPerMinute": {
            "type": "integer"
          },
          "allowedTools": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "blockedTools": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AgentContract": {
        "type": "object",
        "required": [
          "agentId",
          "sla"
        ],
        "properties": {
          "agentId": {
            "type": "string"
          },
          "sla": {
            "$ref": "#/components/schemas/AgentSla"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/contracts/{agentId}": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "Get contract for a specific agent",
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The agent contract",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentContract"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Register or replace a contract for an agent",
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sla"
                ],
                "properties": {
                  "sla": {
                    "$ref": "#/components/schemas/AgentSla"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contract registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentContract"
                }
              }
            }
          },
          "400": {
            "description": "Missing sla field",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Contracts"
        ],
        "summary": "Remove a contract for an agent",
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contract removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contract not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/contracts/{agentId}/check": {
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Pre-check a tool call against an agent's contract (dry run)",
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "toolName"
                ],
                "properties": {
                  "toolName": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pre-check result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "allowed": {
                      "type": "boolean"
                    },
                    "reason": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing toolName",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agents/route": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Route a query to the most relevant agents",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Natural language query to route"
          },
          {
            "in": "query",
            "name": "topN",
            "schema": {
              "type": "integer",
              "default": 3
            },
            "description": "Number of top suggestions to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of agent routing suggestions (sorted by score)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "agentId": {
                        "type": "string"
                      },
                      "score": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}